Move Careerquiz index to pages
